+++ arch/xtensa/core/CMakeLists.txt
@@ -11,6 +11,7 @@ zephyr_library_sources(
   xtensa_asm2_util.S
   xtensa_asm2.c
   irq_manage.c
+  thread.c
   )
 
 zephyr_library_sources_ifdef(CONFIG_XTENSA_USE_CORE_CRT1 crt1.S)
